+++ lib/zstructuredlight/slsystems/stereo/zstereosystemimpl.h
@@ -56,8 +56,8 @@ protected slots:
 protected:
     std::vector<QPointer<Z3D::ZPinholeCameraCalibration> > mCal;
 
-    std::vector< std::vector<cv::Vec3f> > m_undistortedRays;
-    //std::vector< std::vector<cv::Vec3f> > m_undistortedWorldRays;
+    std::vector< std::vector<cv::Vec3d> > m_undistortedRays;
+    //std::vector< std::vector<cv::Vec3d> > m_undistortedWorldRays;
 
     cv::Size m_imageSize;
